refactor: move app queries and migrations out of `internal/db/postgres`.
